Add parser experiment for logfmt (#9)
logfmt is a toy parser for log messages encoded in "logfmt". An
example message is:

	at=info method=GET path=/ host=mutelight.org fwd="124.133.52.161"
	dyno=web.2 connect=4ms service=8ms status=200 bytes=1653

This is a toy because it's slow, and it can panic for some inputs. This
was a learning experiment with github.com/alecthomas/participle/v2.

It would be interesting to compare this implementation to a Ragel one.

References:

	https://www.brandur.org/logfmt
	https://github.com/sirupsen/logrus/blob/f8bf7650dccb756cea26edaf9217aab85500fe07/text_formatter.go