Add fleet-terraform-gen for generate Terraform modules from Fleet packages
fleet-terraform-gen generates a Terraform module in the Terraform JSON syntax
that can be used to manage the installation of single Fleet data stream. It
follows a model of having a single data stream contained in a Fleet package
policy.

It extracts all of the variables definitions from a Fleet package and exposes
them as Terraform variables.

The generated module depends on another module named fleet_package_policy that
handles making the API calls to Fleet to create the policy.