Add akamai-siem-simulator (#26)
Duct-taped simulator for the Akamai SIEM API.

This simulates the Akamai SIEM API. It provides realistic validation of
request signatures. It emulates the time and offset based queries by
simulating a world in which a new event happens every 5 minutes. The
offset cursor is simply the unix time in sec of the last returned event.

**References**

- https://techdocs.akamai.com/siem-integration/reference/get-configid
- https://techdocs.akamai.com/developer/docs/authenticate-with-edgegrid
- https://github.com/akamai/AkamaiOPEN-edgegrid-golang/blob/d417bd104d59eb9bf668da20c35f9bf899b65f90/pkg/edgegrid/signer.go#L44-L54