ecs-update - add smoke test running against elastic/integrations (#34)
Add a smoke test that will run an edit against every package in the
elastic/integrations repo. It will mirror the repo to a local temp dir to
speed up consecutive test runs. Then it performs a clone into a test-local
temp directory.

I added error wrapping to yamlEditString to be able to track down "node not found"
related to packages that have not been updated to remove dotted key names in
the manifest.yml.